修改数据库表及项目名称

+++ src/main/java/com/ahead/school/dao/StudentDao.java
@@ -1,7 +1,8 @@
 package com.ahead.school.dao;
 
-import com.ahead.school.entity.Student;
+
+import com.ahead.school.entity.id;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StudentDao extends JpaRepository<Student, Integer> {
+public interface StudentDao extends JpaRepository<Student,Long>{
 }
